small: bump new version
The main reason of update:
When the memcached module is linked with tarantool,
it uses some of the "small" functions from tarantool
(which were exported by tarantool, for example:
mempool_create_with_order()) and some from the "small"
with wich it was built (for example: mempool_create(),
mempool_free()). This can lead to undefined behavior
(segfault for example). To avoid the problem, will use
the same (close) version of the "small" as tarantool.

Part of #59