memcached: fix stuck on processing GET request
Summary: memcached stuck on processing request when number of
processed commands is exceeded of (batch_count + 1) and buffer has a
request with retrieval command ("get", "gets", "gat", and "gats").

For each incoming connection memcached starts endless loop in function
memcached_loop() where it reads network data to buffer
(memcached_loop_read), parse commands from a buffer one by one
(parse_request), process commands (process_request) and write answers to
network socket (memcached_flush).

Imagine we have a sequence of requests equal to con->cfg->batch_count
(it is a constant value and equal to 20) + 1 with SETQ (SETQ is the same
as a SET command, but it is not required confirmation) commands and a
single GET command that should return a value set by one of the previous
SETQ command. Processing of SETQ commands finished successfully. After
processing the latest request with SETQ command we have a GET command in
a buffer and batch_count counter become equal to 20 (batch_count initial
value is 0). In such case control flow returns to start of the
read-parse-process-write loop where we stuck in memcached_loop_read()
and wait for next requests instead of replying to client with value of
requested key.

    rc = con->cb.process_request(con);
    ...
    if (rc == -1)
            memcached_loop_error(con);
    if (con->close_connection) {
            break;
    } else if (rc == 0 && ibuf_used(con->in) > 0 &&
               batch_count < con->cfg->batch_count) {
            goto next;
    }
    ...
    continue;

Fixes #73