third_party: replace libmemcached Git submodule with binary package
libmemcached is a project that provides a library with implemenation of
memcached protocol and a number of tools for working with memcached
servers. memcached uses memcapable [1] that checks memcached server
implemenation for conformance with memcached protocol. libmemcached
available as binary packages on Ubuntu (memcapable is in
libmemcached-tools, see [2]) and no reasons to keep libmemcached as a Git
submodule and spend time on it's compilation and maintenance.

1. http://docs.libmemcached.org/bin/memcapable.html
2. https://packages.ubuntu.com/focal/libmemcached-tools

Closes #90