memcached: rename internal module function memcached_init()
Name of internal function for creating a new memcached instance is
memcached_init(). It may be confusing - name can be mixed up with
name of function for module initialization.

In the following commit we will need an explicit function
for module initialization, memcached_init() was renamed to
memcached_create_instance() for clarity.

This commit changes internal function's name, no impact on public API.

Needed for #96