memcached: allow to pass port as a number to memcached.create()
For creating memcached instance module requires instance name, URI and
table with options. URI is a string with hostname and port number
separated by colon. Instead URI user can pass a port as a number, but
memcached doesn't expect it and fails with exception. Patch fixes it by
adding conversion of uri to a string.

Fixes #107