ibuf: rename ibuf_*() to memcached_ibuf_*()
See [1] for the problem root and [2] for analysis. In brief: `ibuf_*()`
symbols may clash with the same named symbols in tarantool, so it worth
to rename them to avoid effects of different binary layouts: assertion
fails and memory corruptions.

For example, `ibuf_reserve_slow()` calls `slab_get()` and `slab_put()`.
If something will be changed in tarantool's slab cache, we can meet
problems in the module.

This patch just renames the `ibuf_*()` functions. It is the most simple
and durable workaround of the problem.

It is likely not necessary to rename inline functions, but since I'm not
dead sure, I renamed them all.

It is the last known problem from #59. All others are confirmed as safe
to ignore if we'll keep current libsmall version in the module (see the
analysis in [2]).

[1]: https://github.com/tarantool/tarantool/issues/6873
[2]: https://github.com/tarantool/memcached/issues/59#issuecomment-1081106140

Fixes #59